Rename Feature file to be consistent with other object files.
