+++ requirements-package.txt
@@ -2,3 +2,4 @@ requests
 fcache
 mmh3
 apscheduler
+marshmallow
